refactor: extract figure-range-controls.js from index.js
- figure-range-controls.js: 17 functions (~290 lines) for the collapsible
  'Figure Range' panel — ensurePlotRangeControls(), renderPlotRangeControls(),
  refreshPlotRangeControls() as public API
- Upgraded all internal references from window.GTCGlobal to direct state import
- index.js: 1316 → 849 lines (467 removed across all extractions so far)
