+++ public/javascripts/index.js
@@ -22,6 +22,11 @@ import {
     applyTimeUnitToFigures,
     TIME_UNIT_LABEL,
 } from './units.js';
+import {
+    ensurePlotRangeControls,
+    renderPlotRangeControls,
+    refreshPlotRangeControls,
+} from './figure-range-controls.js';
 
 // status bar on top
 class StatusBar {
@@ -359,295 +364,8 @@ function registerButtons(buttons, cb = getDataThenPlot) {
     });
 }
 
-// refreshTimeUnitFactor, TIME_UNIT_LABEL, and applyTimeUnitToFigures
-// are now imported from units.js
-
-function getVisibleFigureDivs() {
-    return [...document.getElementById('figure-wrapper').children].filter(
-        figure => figure.classList.contains('active') && figure.data?.length
-    );
-}
-
-function getFigureAxisLayout(figure, axisName) {
-    return (
-        figure?._fullLayout?.[`${axisName}axis`] ??
-        figure?.layout?.[`${axisName}axis`]
-    );
-}
-
-function getFigureAxisRange(figure, axisName) {
-    const range = getFigureAxisLayout(figure, axisName)?.range;
-    if (
-        Array.isArray(range) &&
-        range.length === 2 &&
-        range.every(value => Number.isFinite(Number(value)))
-    ) {
-        return range.map(Number);
-    }
-}
-
-function formatRangeValue(value) {
-    const numericValue = Number(value);
-    if (!Number.isFinite(numericValue)) {
-        return '';
-    }
-
-    return String(Number(numericValue.toPrecision(10)));
-}
-
-function getPlotRangeForm(figure) {
-    return window.GTCGlobal.activePanel?.querySelector(
-        `.plot-range-form[data-figure-id="${figure.id}"]`
-    );
-}
-
-function syncFigureRangeControlInputs(figure) {
-    const form = getPlotRangeForm(figure);
-    if (!form) {
-        return;
-    }
-
-    ['x', 'y'].forEach(axisName => {
-        const range = getFigureAxisRange(figure, axisName);
-        const minInput = form.querySelector(
-            `input[data-axis="${axisName}"][data-bound="min"]`
-        );
-        const maxInput = form.querySelector(
-            `input[data-axis="${axisName}"][data-bound="max"]`
-        );
-
-        minInput.value = range ? formatRangeValue(range[0]) : '';
-        maxInput.value = range ? formatRangeValue(range[1]) : '';
-    });
-}
-
-function extractRelayoutRange(eventData, axisName) {
-    const key = `${axisName}axis.range`;
-    const tuple = eventData?.[key];
-    if (
-        Array.isArray(tuple) &&
-        tuple.length === 2 &&
-        tuple.every(value => Number.isFinite(Number(value)))
-    ) {
-        return tuple.map(Number);
-    }
-
-    const minValue = eventData?.[`${key}[0]`];
-    const maxValue = eventData?.[`${key}[1]`];
-    if (Number.isFinite(Number(minValue)) && Number.isFinite(Number(maxValue))) {
-        return [Number(minValue), Number(maxValue)];
-    }
-}
-
-function updateHistoryModeRangeFromRelayout(figure, eventData) {
-    const currentPlotId = window.GTCGlobal.current_plot_btn?.id;
-    if (!currentPlotId?.startsWith('History') || !currentPlotId.includes('-mode')) {
-        return;
-    }
-
-    const targetField =
-        figure.id === 'figure-2'
-            ? 'growthRate'
-            : figure.id === 'figure-3'
-              ? 'frequency'
-              : undefined;
-    if (!targetField) {
-        return;
-    }
-
-    window.GTCGlobal.hist_mode_range[targetField] =
-        extractRelayoutRange(eventData, 'x') ?? getFigureAxisRange(figure, 'x');
-}
-
-function bindFigureRangeSync(figure) {
-    if (figure.dataset.rangeSyncBound === 'true') {
-        return;
-    }
-
-    figure.dataset.rangeSyncBound = 'true';
-    figure.on('plotly_relayout', eventData => {
-        syncFigureRangeControlInputs(figure);
-        updateHistoryModeRangeFromRelayout(figure, eventData);
-    });
-}
-
-function applyFigureRangeFromForm(figure, form) {
-    const updates = {};
-    for (const axisName of ['x', 'y']) {
-        const currentRange = getFigureAxisRange(figure, axisName);
-        const minValue = form.querySelector(
-            `input[data-axis="${axisName}"][data-bound="min"]`
-        ).value;
-        const maxValue = form.querySelector(
-            `input[data-axis="${axisName}"][data-bound="max"]`
-        ).value;
-
-        if (!minValue && !maxValue) {
-            continue;
-        }
-
-        const nextMin = minValue ? Number(minValue) : currentRange?.[0];
-        const nextMax = maxValue ? Number(maxValue) : currentRange?.[1];
-        if (!Number.isFinite(nextMin) || !Number.isFinite(nextMax)) {
-            alert(`Enter valid numeric ${axisName.toUpperCase()} range values.`);
-            return;
-        }
-        if (nextMin >= nextMax) {
-            alert(
-                `${axisName.toUpperCase()} minimum must be smaller than maximum.`
-            );
-            return;
-        }
-
-        updates[`${axisName}axis.range`] = [nextMin, nextMax];
-        updates[`${axisName}axis.autorange`] = false;
-    }
-
-    if (!Object.keys(updates).length) {
-        return;
-    }
-
-    Plotly.relayout(figure, updates).then(() => syncFigureRangeControlInputs(figure));
-}
-
-function resetFigureAutoRange(figure) {
-    Plotly.relayout(figure, {
-        'xaxis.autorange': true,
-        'yaxis.autorange': true,
-    }).then(() => syncFigureRangeControlInputs(figure));
-}
-
-function createPlotRangeField(axisName, bound, value) {
-    const label = document.createElement('label');
-    label.className = 'plot-range-field';
-
-    const text = document.createElement('span');
-    text.innerText = `${axisName.toUpperCase()} ${bound}`;
-
-    const input = document.createElement('input');
-    Object.assign(input, {
-        type: 'text',
-        value: value === undefined ? '' : formatRangeValue(value),
-        placeholder: 'auto',
-        autocomplete: 'off',
-    });
-    input.dataset.axis = axisName;
-    input.dataset.bound = bound;
-
-    label.append(text, input);
-    return label;
-}
-
-function createPlotRangeForm(figure) {
-    const form = document.createElement('form');
-    form.className = 'plot-range-form';
-    form.dataset.figureId = figure.id;
-
-    const title = document.createElement('div');
-    title.className = 'plot-range-title';
-    title.innerText = figure.id.replace('figure-', 'Figure ');
-
-    const grid = document.createElement('div');
-    grid.className = 'plot-range-grid';
-    const xRange = getFigureAxisRange(figure, 'x');
-    const yRange = getFigureAxisRange(figure, 'y');
-    grid.append(
-        createPlotRangeField('x', 'min', xRange?.[0]),
-        createPlotRangeField('x', 'max', xRange?.[1]),
-        createPlotRangeField('y', 'min', yRange?.[0]),
-        createPlotRangeField('y', 'max', yRange?.[1])
-    );
-
-    const actions = document.createElement('div');
-    actions.className = 'plot-range-actions';
-
-    const applyButton = document.createElement('button');
-    applyButton.type = 'submit';
-    applyButton.className = 'tab-l1-btn';
-    applyButton.innerText = 'Apply';
-
-    const autoButton = document.createElement('button');
-    autoButton.type = 'button';
-    autoButton.className = 'tab-l1-btn';
-    autoButton.innerText = 'Auto';
-    autoButton.addEventListener('click', () => resetFigureAutoRange(figure));
-
-    actions.append(applyButton, autoButton);
-    form.append(title, grid, actions);
-    form.addEventListener('submit', event => {
-        event.preventDefault();
-        applyFigureRangeFromForm(figure, form);
-    });
-
-    return form;
-}
-
-function setPlotRangeControlsExpanded(container, expanded) {
-    container.classList.toggle('active', expanded);
-    const toggle = container.querySelector('.plot-range-toggle');
-    toggle.setAttribute('aria-expanded', expanded);
-}
-
-function ensurePlotRangeControls(panel) {
-    let container = panel.querySelector('.plot-range-controls');
-    if (container) {
-        return container;
-    }
-
-    container = document.createElement('div');
-    container.className = 'plot-range-controls';
-
-    const toggle = document.createElement('button');
-    toggle.type = 'button';
-    toggle.className = 'plot-range-toggle';
-    toggle.innerText = 'Figure Range';
-    toggle.addEventListener('click', () => {
-        setPlotRangeControlsExpanded(
-            container,
-            !container.classList.contains('active')
-        );
-    });
-
-    const body = document.createElement('div');
-    body.className = 'plot-range-controls-body';
-
-    container.append(toggle, body);
-    setPlotRangeControlsExpanded(container, false);
-    panel.appendChild(container);
-    return container;
-}
-
-function renderPlotRangeControls(panel = window.GTCGlobal.activePanel) {
-    if (!panel) {
-        return;
-    }
-
-    const body = ensurePlotRangeControls(panel).querySelector(
-        '.plot-range-controls-body'
-    );
-    body.replaceChildren();
-
-    const figures = getVisibleFigureDivs().filter(
-        figure => getFigureAxisLayout(figure, 'x') && getFigureAxisLayout(figure, 'y')
-    );
-    if (!figures.length) {
-        const placeholder = document.createElement('p');
-        placeholder.className = 'plot-range-empty';
-        placeholder.innerText = 'Plot a figure to set its x/y range here.';
-        body.appendChild(placeholder);
-        return;
-    }
-
-    figures.forEach(figure => {
-        bindFigureRangeSync(figure);
-        body.appendChild(createPlotRangeForm(figure));
-    });
-}
-
-function refreshPlotRangeControls() {
-    renderPlotRangeControls(window.GTCGlobal.activePanel);
-    getVisibleFigureDivs().forEach(syncFigureRangeControlInputs);
-}
+// Figure range controls (ensurePlotRangeControls, renderPlotRangeControls,
+// refreshPlotRangeControls) are now imported from figure-range-controls.js
 
 async function openPanel(clean_beforehand = true) {
     if (this.id == 'Summary') {
